Wrote make file

+++ src/cpu_regs.c
@@ -21,9 +21,9 @@ regfile_t *regfile_new(memory_t *M) {
 }
 
 // Prints a regfile.
-void regfile_print(regfile_t *R, int i) {
+void regfile_print(regfile_t *R, size_t i) {
   printf("-----------------------------------\n");
-  printf("State following iteration %d:\n", i);
+  printf("State following iteration %d:\n", (int)i);
   printf("A: %x, X: %x, Y: %x, INST: %x\n", R->A, R->X, R->Y, R->inst);
   printf("State (flags): %x, Stack pointer: %x\n", R->P, R->S);
   printf("PCL: %x, PCH: %x\n", R->PCL, R->PCH);
